renamed gamma main component
